+++ core/src/main/java/io/cucumber/core/cli/Main.java
@@ -74,8 +74,7 @@ public class Main {
                 .parse(argv)
                 .addDefaultGlueIfAbsent()
                 .addDefaultFeaturePathIfAbsent()
-                .addDefaultFormatterIfAbsent()
-                .addDefaultSummaryPrinterIfAbsent()
+                .addDefaultSummaryPrinterIfNotDisabled()
                 .enablePublishPlugin()
                 .build(systemOptions);